Split database writing into a separate app.
`reader` now reads from the TCP publisher, parses the messages as JSON and then publishes them to a RabbitMQ queue,
`consumer` consumes messages from the RabbitMQ queue and persists each one in the database.

Added log levels and terminal colours to logger and moved it to `library`.
Added RabbitMQ to `docker-compose.yml`.